remove contacts and rename to preferences
