+++ library/src/main/java/org/xmtp/android/library/Group.kt
@@ -151,15 +151,7 @@ class Group(val client: Client, private val libXMTPGroup: FfiConversation) {
         return Message(client, message)
     }
 
-    suspend fun updateConsentState(state: ConsentState) {
-        if (client.hasV2Client) {
-            when (state) {
-                ConsentState.ALLOWED -> client.contacts.allowGroups(groupIds = listOf(id))
-                ConsentState.DENIED -> client.contacts.denyGroups(groupIds = listOf(id))
-                ConsentState.UNKNOWN -> Unit
-            }
-        }
-
+    fun updateConsentState(state: ConsentState) {
         val consentState = ConsentState.toFfiConsentState(state)
         libXMTPGroup.updateConsentState(consentState)
     }
